[Feat] UIManager
